refactor(#741): refactor languageController
