--- api/controllers/v1/LanguageController.js
@@ -1,9 +1,0 @@
-/**
- */
-
-const languageController = require('../LanguageController');
-
-module.exports = {
-  find: (req, res) => languageController.find(req, res),
-  findAll: (req, res) => languageController.findAll(req, res),
-};
